Minor extensions to rw::common::DOMElem Fixed warning in rw::common::IOUtil
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@4898 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rw/common/IOUtil.cpp
@@ -82,7 +82,8 @@ void IOUtil::readFile(
 
     // The file length.
     inp.seekg(0, std::ios::end);
-    const int len = inp.tellg();
+    const int len = (int)inp.tellg();
+	
     inp.seekg(0, std::ios::beg);
 
     // The buffer.
